Introduce ``sphinx.ext.autodoc._dynamic`` subpackage
