+++ sphinx/ext/autodoc/_dynamic/__init__.py
@@ -1,0 +1,1 @@
+"""The dynamic (import-based) backend for autodoc."""
