+++ sphinx/ext/autodoc/_dynamic/_signatures.py
@@ -7,11 +7,13 @@ from inspect import Parameter, Signature
 from typing import TYPE_CHECKING, NewType, TypeVar
 
 from sphinx.errors import PycodeError
+from sphinx.ext.autodoc._dynamic._preserve_defaults import update_default_value
+from sphinx.ext.autodoc._dynamic._type_annotations import _record_typehints
+from sphinx.ext.autodoc._dynamic._type_comments import (
+    _update_annotations_using_type_comments,
+)
 from sphinx.ext.autodoc._names import py_ext_sig_re
 from sphinx.ext.autodoc._property_types import _AssignStatementProperties
-from sphinx.ext.autodoc._type_comments import _update_annotations_using_type_comments
-from sphinx.ext.autodoc.preserve_defaults import update_default_value
-from sphinx.ext.autodoc.typehints import _record_typehints
 from sphinx.locale import __
 from sphinx.pycode import ModuleAnalyzer
 from sphinx.util import inspect, logging
