+++ sphinx/ext/autosummary/__init__.py
@@ -67,12 +67,12 @@ import sphinx
 from sphinx import addnodes
 from sphinx.errors import PycodeError
 from sphinx.ext.autodoc._directive_options import _AutoDocumenterOptions
-from sphinx.ext.autodoc._importer import _import_module
-from sphinx.ext.autodoc._loader import _load_object_by_name
-from sphinx.ext.autodoc._member_finder import _best_object_type_for_member
+from sphinx.ext.autodoc._dynamic._importer import _import_module
+from sphinx.ext.autodoc._dynamic._loader import _load_object_by_name
+from sphinx.ext.autodoc._dynamic._member_finder import _best_object_type_for_member
+from sphinx.ext.autodoc._dynamic._mock import mock
 from sphinx.ext.autodoc._sentinels import INSTANCE_ATTR
 from sphinx.ext.autodoc._shared import _AutodocAttrGetter, _AutodocConfig
-from sphinx.ext.autodoc.mock import mock
 from sphinx.locale import __
 from sphinx.pycode import ModuleAnalyzer
 from sphinx.util import logging, rst
